shuffle bits around
