+++ src/util/peek_while.rs
